requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.2
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
xlrd==2.0.2
xlsxwriter==3.2.9
